% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_col_if_absent}
\alias{add_col_if_absent}
\title{Add Column if Absent}
\usage{
add_col_if_absent(data, column_name, na_type = NA)
}
\arguments{
\item{data}{a data.frame or tibble to which a column may be added}

\item{column_name}{the name of the column to be added if it doesn't exist}

\item{na_type}{the type of NA value to use for the column values. Default to basic \code{NA}}
}
\value{
the data.frame as provided (converted to tibble)
}
\description{
Adds a column (with the name specified in column_name) of NA values to a data.frame or tibble. If that
column already exists, no change will be made to data. NA value type (character, integer, real, logical)
may be specified.
}
\keyword{internal}
