% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ff1_session_loaded}
\alias{check_ff1_session_loaded}
\title{Check FastF1 Session Loaded}
\usage{
check_ff1_session_loaded(session_name = "session")
}
\arguments{
\item{session_name}{Name of the python session object. For internal functions, typically \code{session}.}
}
\value{
invisible TRUE, no real return, called for effect
}
\description{
Used to verify that the fastf1 session is loaded before trying to work with it.

Prevents errors in automated processing code.
}
\keyword{internal}
