% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ff1_version}
\alias{check_ff1_version}
\title{Check FastF1 Version}
\usage{
check_ff1_version()
}
\value{
Invisibly \code{TRUE} if not raising an error for unsupported \code{FastF1} version.
}
\description{
This function checks the version of \code{FastF1} and ensures it's at or above the minimum supported version for
\code{f1dataR} (currently requires 3.4.0 or better).

This function is a light wrapper around get_fastf1_version()
}
\keyword{internal}
