% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_f1_cache.R
\name{clear_f1_cache}
\alias{clear_f1_cache}
\title{Clear f1fastR Cache}
\usage{
clear_f1_cache()
}
\value{
No return value, called to erase cached data
}
\description{
Clears the cache for f1dataR telemetry and Ergast API results.
Note that the cache directory can be set by setting \verb{option(f1dataR.cache = [cache dir])},
but the default is but the default is a temporary directory.
}
\examples{
\dontrun{
clear_f1_cache()
}
}
