% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constructor_data}
\alias{constructor_data}
\title{Constructor Data}
\description{
This contains the following columns:
constructor_id, constructor_color, constructor_color2, constructor_logo
}
\references{
derived from fastf1 data
}
\keyword{data}
