% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_circuits.R
\name{.load_circuits}
\alias{.load_circuits}
\title{Load Circuit Info (not cached)}
\usage{
.load_circuits(season = "current")
}
\arguments{
\item{season}{number from 1950 to current season (defaults to current season).}
}
\value{
A dataframe with columns circuitID, circuitName, Location latitude
and longitude, locality (usually city) and country.
}
\description{
Loads circuit info for all circuits in a given season.
This function does not export, only the cached version.
}
