% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_drivers.R
\name{.load_drivers}
\alias{.load_drivers}
\title{Load Driver Info (not cached)}
\usage{
.load_drivers(season = "current")
}
\arguments{
\item{season}{number from 1950 to current season (defaults to current season).}
}
\value{
A tibble with one row per driver.
}
\description{
Loads driver info for all participants in a given season.
This function does not export, only the cached version.
}
