% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_laps.R
\name{.load_laps}
\alias{.load_laps}
\title{Load Lap by Lap Time Data (not cached)}
\usage{
.load_laps(season = "current", race = "last")
}
\arguments{
\item{season}{number from 1996 to current season (defaults to current season)}

\item{race}{number from 1 to 23 (depending on season selected) and defaults
to most recent}
}
\value{
A dataframe with columns driverId (unique and recurring), position
during lap, time (in clock form), lap number, time (in seconds), and season.
}
\description{
Loads lap-by-lap time data for all drivers in a given season
and round. Lap time data is available from 1996 onward. This function does not export, only the cached version.
}
