% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_results.R
\name{.load_results}
\alias{.load_results}
\title{Load Results (not cached)}
\usage{
.load_results(season = "current", round = "last")
}
\arguments{
\item{season}{number from 1950 to 2022 (defaults to current season).}

\item{round}{number from 1 to 23 (depending on season), and defaults
to most recent.}
}
\value{
A dataframe with columns driverId, grid position, laps completed,
race status (finished or otherwise), gap to first place, fastest lap rank,
fastest lap time, fastest lap in seconds, and top speed in kph.
}
\description{
Loads final race resuts for a given year and round.
}
