% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_schedule.R
\name{.load_schedule}
\alias{.load_schedule}
\title{Load Schedule (not cached)}
\usage{
.load_schedule(season = "current")
}
\arguments{
\item{season}{number from 1950 to current season (defaults to current season).}
}
\value{
A tibble with one row per circuit in season
}
\description{
Loads schedule information for a given F1 season.
This function does not export, only the cached version.
}
