% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_standings.R
\name{.load_standings}
\alias{.load_standings}
\title{Load Standings (not cached)}
\usage{
.load_standings(season = "current", round = "last", type = "driver")
}
\arguments{
\item{season}{number from 1950 to current season (defaults to current season).}

\item{round}{number from 1 to 23 (depending on season), and defaults
to most recent.}

\item{type}{select drivers or constructors championship data. Defaults to
drivers}
}
\value{
A tibble with one row per driver.
}
\description{
Loads standings at the end of a given season and round for drivers' or
constructors' championships.
}
