% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{driver_constructor_data}
\alias{driver_constructor_data}
\title{Driver/Constructor Data}
\description{
This contains the following columns:
driverId, fullName, driverCode, constructorId, constructorName, constructorColor, constructorColor2, constructorLogo, driverNumber, numberImage, season
The data is derived from fastf1 python package and can be updated by
running \code{f1dataR:::build_driver_constructor_data}
}
\references{
derived from fastf1 data
}
\keyword{data}
