% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_current_season}
\alias{get_current_season}
\title{Get Current Season core}
\usage{
get_current_season()
}
\value{
Year (four digit number) representation of current season, as numeric.
}
\description{
Looks up current season from ergast, fallback to manual determination
}
