% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_driver_telemetry.R
\name{get_driver_telemtery}
\alias{get_driver_telemtery}
\title{Load Telemetry Data for a Driver}
\usage{
get_driver_telemtery(
  season = get_current_season(),
  round = 1,
  session = "R",
  driver,
  fastest_only = FALSE,
  log_level = "WARNING",
  race = lifecycle::deprecated()
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{load_driver_telemtery()} was renamed to \code{load_driver_telemetry()} to create a more
consistent API.
}
\details{
Receives season, race number, driver code, and an optional fastest lap only
argument to output car telemetry for the selected situation.
}
\keyword{internal}
