% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ergast_content}
\alias{get_ergast_content}
\title{Get Ergast Content}
\usage{
get_ergast_content(url)
}
\arguments{
\item{url}{the Ergast URL tail to get from the API (for example,
\code{"{season}/circuits.json?limit=40"} is called from \code{load_circuits()}).}
}
\value{
the result of \code{jsonlite::fromJSON} called on Ergast's return content.
Further processing is performed by specific functions
}
\description{
Gets Ergast content and returns the processed json object if
no Ergast errors are found. This will automatically fall back from https://
to http:// if Ergast suffers errors, and will automatically retry up to 10
times by each protocol

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Note the Ergast Motor Racing Database API shut down at the end of 2024.
This function was replaced by the Jolpica data source.
}
\keyword{internal}
