% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{get_tire_compounds}
\alias{get_tire_compounds}
\title{Get Tire Compounds}
\usage{
get_tire_compounds(season = get_current_season())
}
\arguments{
\item{season}{number from 2018 to current season. Defaults to current season.}
}
\value{
a data.frame with two columns: \code{compound} and \code{color}
}
\description{
Get a data.frame of all tire compound names and associated colours for a season.
}
\examples{
if (interactive()) {
  # To get this season's tires
  get_tire_compounds()

  # Compare to 2018 tires:
  get_tire_compounds(2018)
}
}
