% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_circuits.R
\name{load_circuits}
\alias{load_circuits}
\title{Load Circuit Info}
\usage{
load_circuits(season = get_current_season())
}
\arguments{
\item{season}{number from 1950 to current season, or \code{"current"} (defaults to current season).}
}
\value{
A tibble with one row per circuit. Returns \code{NULL} on API failure.
}
\description{
Loads circuit info for all circuits in a given season.
}
