% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_race_laps.R
\name{load_race_laps}
\alias{load_race_laps}
\title{Load Lapwise Data}
\usage{
load_race_laps(
  obj_name = "session_laps",
  season = get_current_season(),
  race = 1,
  session = "R",
  log_level = "WARNING",
  add_weather = F
)
}
\arguments{
\item{obj_name}{name assigned to the loaded session to be referenced later.}

\item{season}{number from 2018 to current season. Defaults to current season}

\item{race}{number from 1 to 23 (depending on season selected) and defaults
to most recent. Also accepts race name.}

\item{session}{the code for the session to load Options are FP1, FP2, FP3,
Q, S, and R. Default is "R", which refers to Race.
Cache directory can be set by setting \verb{option(f1dataR.cache = [cache dir])},
default is the current working directory.}

\item{log_level}{Detail of logging from fastf1 to be displayed. Choice of:
DEBUG, INFO, WARNING, ERROR and CRITICAL. See \href{https://theoehrly.github.io/Fast-F1/fastf1.html#configure-logging-verbosity}{fastf1 documentation}.}

\item{add_weather}{Whether to add weather information to the laps. See \href{https://docs.fastf1.dev/core.html#fastf1.core.Laps.get_weather_data}{fastf1 documentation} for info on weather}
}
\value{
A data frame. Note time information is in seconds, see \href{https://docs.fastf1.dev/time_explanation.html}{fastf1 documentation} for more information on timing.
}
\description{
Loads lapwise data for a race session.
}
\details{
Includes each driver's each lap's laptime, pit in/out time, tyre information, track status, and (optionally) weather information.
}
