% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_race_session.R
\name{load_race_session}
\alias{load_race_session}
\title{Load Session Data}
\usage{
load_race_session(
  obj_name = "session",
  season = get_current_season(),
  round = 1,
  session = "R",
  log_level = "WARNING",
  race = lifecycle::deprecated()
)
}
\arguments{
\item{obj_name}{name assigned to the loaded session to be referenced later.
Leave as \code{'session'} unless otherwise required.}

\item{season}{number from 2018 to current season. Defaults to current season.}

\item{round}{number from 1 to 24 (depending on season selected), defaults
to \code{1}. Also accepts race name.}

\item{session}{the code for the session to load. Options are \code{'FP1'}, \code{'FP2'}, \code{'FP3'},
\code{'Q'}, \code{'S'}, \code{'SS'},\code{'SQ'}, and \code{'R'}. Default is \code{'R'}, which refers to Race.}

\item{log_level}{Detail of logging from fastf1 to be displayed. Choice of:
\code{'DEBUG'}, \code{'INFO'}, \code{'WARNING'}, \code{'ERROR'} and \code{'CRITICAL'}. See
\href{https://docs.fastf1.dev/fastf1.html#configure-logging-verbosity}{fastf1 documentation}.}

\item{race}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{race} is no longer supported, use \code{round}}
}
\value{
Invisibly, a \code{reticulate} Python FastF1 session object, primarily useful
for advanced users who want to run their own \code{reticulate::py_run_string()} calls
against it. Returns \code{NULL} if the session fails to load.
}
\description{
Loads telemetry and general data from the official F1 data stream via
the fastf1 python library. Data is available from 2018 onward.

The data loaded can optionally be assigned to a R variable, and then interrogated for
session data streams. See the \href{https://docs.fastf1.dev/}{fastf1 documentation}
for more details on the data returned by the python API.

Cache directory can be set by setting \verb{options(f1dataR.cache = [cache dir])}.
The default option is \code{"memory"}; when the option is \code{"memory"}, \code{"off"}, or
\code{"filesystem"}, the underlying FastF1 HTTP cache is written to \code{tempdir()}.
}
\examples{
# Load the quali session from 2019 first round
if (interactive()) {
  session <- load_race_session(season = 2019, round = 1, session = "Q")
}
}
\seealso{
\code{\link[=load_session_laps]{load_session_laps()}}, \code{\link[=plot_fastest]{plot_fastest()}}
}
