% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_schedule.R
\name{load_schedule}
\alias{load_schedule}
\title{Load Schedule}
\usage{
load_schedule(season = get_current_season())
}
\arguments{
\item{season}{number from 1950 to current season (defaults to current season). \code{'current'} also accepted.}
}
\value{
A tibble with one row per round in season. Indicates in sprint_date if a specific round has a sprint race.
Returns \code{NULL} on API failure.
}
\description{
Loads schedule information for a given F1 season.
}
