% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sprint.R
\name{load_sprint}
\alias{load_sprint}
\title{Load Sprint Results}
\usage{
load_sprint(season = get_current_season(), round = "last")
}
\arguments{
\item{season}{number from 2021 to current season (or the word 'current') (defaults to current season).}

\item{round}{number from 1 to the number of rounds in the season, and defaults
to most recent. Also accepts \code{'last'}.}
}
\value{
A tibble with columns driver_id, constructor_id, points awarded, finishing position,
grid position, laps completed, race status (finished or otherwise), gap to
first place, fastest lap's lap number, fastest lap time, fastest lap in seconds,
or NULL if no sprint exists for this season/round combo. Results are paginated
automatically, so sprints with more than 100 results are returned in full. Note that
\code{points}, \code{position}, \code{grid}, and \code{laps} are returned by the API as strings and are
therefore all \strong{character} columns, not numeric.
}
\description{
Loads final sprint race results for a given year and round. Note not all rounds have
sprint results.
}
