% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_to_sec}
\alias{time_to_sec}
\title{Convert Clock time to seconds}
\usage{
time_to_sec(time)
}
\arguments{
\item{time}{character string with clock format (0:00.000)}
}
\value{
A numeric variable that represents that time in seconds
}
\description{
This function converts clock format time (0:00.000) to seconds (0.000s)
}
\keyword{internal}
