% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_f1_cache.R
\name{clear_f1_cache}
\alias{clear_f1_cache}
\alias{clear_cache}
\title{Clear f1dataR Cache}
\usage{
clear_f1_cache()

clear_cache()
}
\value{
No return value, called to erase cached data
}
\description{
Clears the cache for f1dataR telemetry and Jolpica API results.
Note that the cache directory can be set by setting \verb{option(f1dataR.cache = [cache dir])},
but the default is a temporary directory.

You can also call the alias \code{clear_cache()} for the same result
}
\examples{
\dontrun{
clear_f1_cache()
}
}
