% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fastest.R
\name{correct_track_ratio}
\alias{correct_track_ratio}
\title{Correct Track Ratios}
\usage{
correct_track_ratio(trackplot, x = "x", y = "y")
}
\arguments{
\item{x, y}{Names of columns in the original data used for the plot's x and y values.
Defaults to 'x' and 'y'}

\item{plot}{A GGPlot object.}
}
\value{
a ggplot object with \code{\link[ggplot2]{scale_x_continuous}} and \code{\link[ggplot2]{scale_y_continuous}} set to the
same limits to produce an image with shared x and y limits and with \code{\link[ggplot2]{coord_fixed}}
}
\description{
Correct Track Ratios helps ensure that ggplot objects are plotted with 1:1 unit ratio.
Without this function, plots have different x & y ratios and the tracks come out misshapen.
This is particularly evident at long tracks like Saudi Arabia or Canada.

Note that this leaves the plot object on a dark background, any plot borders will be maintained
}
\examples{
\dontrun{
# Note that plot_fastest plots have already been ratio corrected
fast_plot <- plot_fastest(season = 2022, round = 1, session = "Q", driver = V)
correct_track_ratio(fast_plot)
}
}
