% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_constructors.R
\name{.load_constructors}
\alias{.load_constructors}
\title{Load Constructor Info (not cached)}
\usage{
.load_constructors()
}
\value{
A dataframe with columns constructorId (unique and recurring),
name, and nationality
}
\description{
Loads constructor info for all participants in a given season.
This function does not export, only the cached version.
}
