% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_drivers.R
\name{.load_drivers}
\alias{.load_drivers}
\title{Load Driver Info (not cached)}
\usage{
.load_drivers(season = 2022)
}
\arguments{
\item{season}{number from 1950 to 2022 (defaults to current season). All
drivers after 2014 will have a permanent number.}
}
\value{
A dataframe with columns driverId (unique and recurring), first name,
last name, nationality, date of birth (yyyy-mm-dd format), driver code, and
permanent number (for post-2014 drivers).
}
\description{
Loads driver info for all participants in a given season.
This funtion does not export, only the cached version.
}
