% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_quali.R
\name{.load_quali}
\alias{.load_quali}
\title{Load Qualifying Results (not cached)}
\usage{
.load_quali(season = "current", round = "last")
}
\arguments{
\item{season}{number from 1950 to current season (defaults to current season).}

\item{round}{number from 1 to 23 (depending on season), and defaults
to most recent.}
}
\value{
A dataframe with columns driverId, obtained position, Q1, Q2, and Q3
times in clock format as well as seconds.
}
\description{
Loads qualifying session results for a given season and round.
}
