% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_standings.R
\name{.load_standings}
\alias{.load_standings}
\title{Load Qualifying Results (not cached)}
\usage{
.load_standings(season = "current", round = "last", type = "driver")
}
\arguments{
\item{season}{number from 1950 to 2022 (defaults to current season).}

\item{round}{number from 1 to 23 (depending on season), and defaults
to most recent.}

\item{type}{select drivers or constructors championship data. Defaluts to
drivers}
}
\value{
A dataframe with columns driverId (or constructorId), position,
points, wins and constructorsId in the case of drivers championship.
}
\description{
Loads qualifying session results for a given season and round.
}
