% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_driver_telemetry.R
\name{get_driver_telemetry}
\alias{get_driver_telemetry}
\title{Load Telemetry Data for a Driver}
\usage{
get_driver_telemetry(session_name, driver, fastest_only = FALSE)
}
\arguments{
\item{session_name}{the name assigned to a session object
(generated by load_race_session)}

\item{driver}{three letter driver code (see load_drivers() for a list)}

\item{fastest_only}{boolean value whether to pick all laps or only the fastest
by the driver in that session.}
}
\value{
A dataframe with columns driverId, lap, stop (number), time (of day),
and stop duration
}
\description{
Recieves a session object name (from load_race_session()) and a
dirver code. Outputs telemetry data for that driver during that session.
}
