% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_driver_telemetry.R
\name{get_driver_telemetry}
\alias{get_driver_telemetry}
\title{Load Telemetry Data for a Driver}
\usage{
get_driver_telemetry(season, race, driver, fastest_only = FALSE)
}
\arguments{
\item{season}{number from 2018 to 2022 (defaults to current season).}

\item{race}{number from 1 to 23 (depending on season selected) and defaults
to most recent. Also accepts race name.}

\item{driver}{three letter driver code (see load_drivers() for a list)}

\item{fastest_only}{boolean value whether to pick all laps or only the fastest
by the driver in that session.}
}
\value{
A dataframe with columns driverId, lap, stop (number), time (of day),
and stop duration
}
\description{
Recieves season, race number, driver code, and an optional fastest lap only
argument to output car telemetry for the selected situation.
}
