% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_driver_telemetry.R
\name{get_driver_telemetry}
\alias{get_driver_telemetry}
\title{Load Telemetry Data for a Driver}
\usage{
get_driver_telemetry(
  season = "current",
  race = "last",
  session = "R",
  driver,
  fastest_only = FALSE
)
}
\arguments{
\item{season}{number from 2018 to 2022 (defaults to current season).}

\item{race}{number from 1 to 23 (depending on season selected) and defaults
to most recent. Also accepts race name.}

\item{session}{the code for the session to load Options are FP1, FP2, FP3,
Q, and R. DEfault is "R", which refers to Race.}

\item{driver}{three letter driver code (see load_drivers() for a list)}

\item{fastest_only}{boolean value whether to pick all laps or only the fastest
by the driver in that session.}
}
\value{
A dataframe with telemetry data for selected driver/session.
}
\description{
Recieves season, race number, driver code, and an optional fastest lap only
argument to output car telemetry for the selected situation.
}
