% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ergast_content}
\alias{get_ergast_content}
\title{Get Ergast Content}
\usage{
get_ergast_content(url)
}
\arguments{
\item{url}{the complete Ergast URL to get}
}
\value{
the result of \code{jsonlite::fromJSON} called on ergast's return content
}
\description{
Gets ergast content and returns the processed json object if
no Ergast errors are found.
}
