% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fastf1_version}
\alias{get_fastf1_version}
\title{Get current FastF1 version}
\usage{
get_fastf1_version()
}
\value{
integer for major version number (or NA if any error )
}
\description{
Gets the current installed FastF1 version available (via \code{reticulate}) to the function.
Displays a note if significantly out of date.
}
