% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_circuits.R
\name{load_circuits}
\alias{load_circuits}
\title{Load Circuit Info}
\usage{
load_circuits(season = get_current_season())
}
\arguments{
\item{season}{number from 1950 to current season (defaults to current season).}
}
\value{
A tibble with one row per circuit
}
\description{
Loads circuit info for all circuits in a given season. Use \code{.load_circuits()}
for an uncached version of this function
}
\keyword{internal}
