% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_constructors.R
\name{load_constructors}
\alias{load_constructors}
\title{Load Constructor Info}
\usage{
load_constructors()
}
\value{
A tibble with one row per constructor
}
\description{
Loads constructor info for all participants in a given season. Use \code{.load_constructors()}for an uncached version of this function
}
\examples{
# Load the list of all constructors who have participated in a F1 race:
load_constructors()
}
