% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_driver_telemetry.R
\name{load_driver_telemetry}
\alias{load_driver_telemetry}
\title{Load Telemetry Data for a Driver}
\usage{
load_driver_telemetry(
  season = get_current_season(),
  round = 1,
  session = "R",
  driver,
  fastest_only = FALSE,
  log_level = "WARNING",
  race = lifecycle::deprecated()
)
}
\arguments{
\item{season}{number from 2018 to current season (defaults to current season).}

\item{round}{number from 1 to 23 (depending on season selected). Also accepts race name.}

\item{session}{the code for the session to load Options are \code{'FP1'}, \code{'FP2'}, \code{'FP3'},
\code{'Q'}, \code{'S'}, \code{'SS'}, and \code{'R'}. Default is \code{'R'}, which refers to Race.}

\item{driver}{three letter driver code (see load_drivers() for a list).}

\item{fastest_only}{boolean value whether to pick all laps or only the fastest
by the driver in that session.}

\item{log_level}{Detail of logging from fastf1 to be displayed. Choice of:
\code{'DEBUG'}, \code{'INFO'}, \code{'WARNING'}, \code{'ERROR'} and \code{'CRITICAL'}. See \href{https://theoehrly.github.io/Fast-F1/fastf1.html#configure-logging-verbosity}{fastf1 documentation}.}

\item{race}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{race} is no longer supported, use \code{round}.}
}
\value{
A tibble with telemetry data for selected driver/session.
}
\description{
Receives season, race number, driver code, and an optional fastest lap only
argument to output car telemetry for the selected situation.
}
\examples{
\dontrun{
telem <- load_driver_telemetry(season = 2023,
                               round = 'bahrain', s
                               ession = 'Q',
                               driver = 'HAM',
                               fastest_only = TRUE)
}

}
