% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_drivers.R
\name{load_drivers}
\alias{load_drivers}
\title{Load Driver Info}
\usage{
load_drivers(season = get_current_season())
}
\arguments{
\item{season}{number from 1950 to current season, or \code{"current"} (defaults to current season).}
}
\value{
A tibble with columns \code{driver_id} (unique and recurring), \code{given_name},
\code{family_name}, \code{nationality}, \code{date_of_birth} (yyyy-mm-dd format), \code{code} (driver code), and
\code{permanent_number} (for post-2014 drivers). Returns \code{NULL} on API failure.
}
\description{
Loads driver info for all participants in a given season.
}
