% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_drivers.R
\name{load_drivers}
\alias{load_drivers}
\title{Load Driver Info}
\usage{
load_drivers(season = get_current_season())
}
\arguments{
\item{season}{number from 1950 to current season (defaults to current season).}
}
\value{
A tibble with columns driver_id (unique and recurring), first name,
last name, nationality, date of birth (yyyy-mm-dd format), driver code, and
permanent number (for post-2014 drivers).
}
\description{
Loads driver info for all participants in a given season. Use \code{.load_drivers()} for an uncached version of this function.
}
\examples{
#Load the drivers appearing in this season's data:
load_drivers()

#Load the drivers who participated in the first year of Formula 1:
load_drivers(1950)
}
