% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pitstops.R
\name{load_pitstops}
\alias{load_pitstops}
\title{Load Pitstop Data}
\usage{
load_pitstops(
  season = get_current_season(),
  round = "last",
  race = lifecycle::deprecated()
)
}
\arguments{
\item{season}{number from 2011 to current season (defaults to current season).}

\item{round}{number from 1 to the number of rounds in the season, and defaults
to most recent. Also accepts \code{'last'}.}

\item{race}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{race} is no longer supported, please use \code{round}.}
}
\value{
A tibble with columns driver_id, lap, stop (number), time (of day),
and stop duration, or NULL if the request fails. Results are paginated automatically,
so races with more than 100 pit stops are returned in full. Note that \code{lap}, \code{stop},
and \code{duration} are returned by the API as strings and are therefore all \strong{character}
columns, not numeric.
}
\description{
Loads pit stop info (number, lap, time elapsed) for a given race
in a season. Pit stop data is available from 2011 onward.
}
