% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_quali.R
\name{load_quali}
\alias{load_quali}
\title{Load Qualifying Results}
\usage{
load_quali(season = get_current_season(), round = "last")
}
\arguments{
\item{season}{number from 2003 to current season (defaults to current season).}

\item{round}{number from 1 to 23 (depending on season), and defaults
to most recent.}
}
\value{
A tibble with one row per driver
}
\description{
Loads qualifying session results for a given season and round. Use \code{.load_quali()} for an uncached version.
}
\examples{
# Load quali from the first race of 2023:
load_quali(2023, 1)

# Load quali from the third race of 2004:
load_quali(2004, 1)
}
