% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_race_session.R
\name{load_race_session}
\alias{load_race_session}
\title{Load Session Data}
\usage{
load_race_session(obj_name, season = 2022, race = 1, session = "R", cache = T)
}
\arguments{
\item{season}{number from 2018 to 2022 (defaults to current season).}

\item{race}{number from 1 to 23 (depending on season selected) and defaults
to most recent. Also accepts race name.}
}
\value{
A session object to be used in other functions.
}
\description{
Loads telemetry and general data from the official F1
data stream via the fastf1 python library. Data is available from
2018 onward.
}
