% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_results.R
\name{load_results}
\alias{load_results}
\title{Load Results}
\usage{
load_results(season = "current", round = "last")
}
\arguments{
\item{season}{number from 1950 to current season (defaults to current season).}

\item{round}{number from 1 to 23 (depending on season), and defaults
to most recent.}
}
\value{
A tibble with one row per driver
}
\description{
Loads final race results for a given year and round.
}
