% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_results.R
\name{load_results}
\alias{load_results}
\title{Load Results}
\usage{
load_results(season = get_current_season(), round = "last")
}
\arguments{
\item{season}{number from 1950 to current season (or the word 'current') (defaults to current season).}

\item{round}{number from 1 to 23 (depending on season), and defaults
to most recent. Also accepts \code{'last'}.}
}
\value{
A tibble with one row per driver
}
\description{
Loads final race results for a given year and round. Use \code{.load_results()} for an uncached version
}
\examples{
# Load results from the third race of 2023
load_results(2023, 3)

# Load results from the first race of 1995
load_results(1995, 1)

# Load results from the last race of 1963
load_results(1963, "last")

}
