% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_schedule.R
\name{load_schedule}
\alias{load_schedule}
\title{Load Schedule}
\usage{
load_schedule(season = "current")
}
\arguments{
\item{season}{number from 1950 to current season (defaults to current season).}
}
\value{
A dataframe with columns season, round, circuitId, circuitName,
latitute and Longitude, locality (city usually), country, date, and time
of the race.
}
\description{
Loads schedule information for a given F1 season.
}
