% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sprint.R
\name{load_sprint}
\alias{load_sprint}
\title{Load Sprint}
\usage{
load_sprint(season = "current", round = "last")
}
\arguments{
\item{season}{number from 2021 to current season (defaults to current season).}

\item{round}{number from 1 to 23 (depending on season), and defaults
to most recent.}
}
\value{
A dataframe with columns driverId, points awarded, finishing position,
grid position, laps completed, race status (finished or otherwise), gap to
first place, fastest lap, fastest lap time, fastest lap in seconds,
or NULL if no sprint exists for this season/round combo
}
\description{
Loads sprint race results for a given year and round. Note not
all rounds have sprint results
}
