% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fastest.R
\name{plot_fastest}
\alias{plot_fastest}
\title{Load Session Data}
\usage{
plot_fastest(season, race, driver, color = "gear")
}
\arguments{
\item{season}{number from 2018 to 2022 (defaults to current season).}

\item{race}{number from 1 to 23 (depending on season selected) and defaults
to most recent.}

\item{driver}{three letter driver code (see load_drivers() for a list)}

\item{color}{argument that indicates which variable to plot overtop the
circuit}
}
\value{
A ggplot object that indicates grand prix, driver, time and selected
color variable.
}
\description{
Loads telemetry and general data from the official F1
data stream via the fastf1 python library. Data is available from
2018 onward.
}
