% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fastest.R
\name{plot_fastest}
\alias{plot_fastest}
\title{Plot Fastest Lap}
\usage{
plot_fastest(season = 2022, race = 1, session = "R", driver, color = "gear")
}
\arguments{
\item{season}{number from 2018 to current season (defaults to current season).}

\item{race}{number from 1 to 23 (depending on season selected) and defaults
to most recent.}

\item{session}{the code for the session to load Options are FP1, FP2, FP3,
Q, S, SS, and R. Default is "R", which refers to Race.}

\item{driver}{three letter driver code (see load_drivers() for a list)}

\item{color}{argument that indicates which variable to plot overtop the
circuit}
}
\value{
A ggplot object that indicates grand prix, driver, time and selected
color variable.
}
\description{
Creates a ggplot object that details the fastest lap for a driver in a race.
Complete with a gearshift or speed analysis.
}
