% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_fastf1}
\alias{setup_fastf1}
\title{Setup fastf1 connection}
\usage{
setup_fastf1(envname = "f1dataRenv", conda = FALSE)
}
\arguments{
\item{envname}{a name for the virtualenv or conda environment.
For virtualenv, if a name is passed, \code{reticulate} will use/create the environment in the
default location. Alternatively, if providing a full path, \code{reticulate} will use the specified location.}

\item{conda}{whether to use conda environments or virtualenvs. Default FALSE (i.e. virtualenv)}
}
\value{
No return value, called to set up virtan enviroment for package
}
\description{
Set up reticulate using some options from user (or defaults). Helps
solve \code{fastf1} issues - see the Setup FastF1 Connection vignette for more info
(run \code{vignette('setup_fastf1', 'f1dataR')}).
}
\examples{
\dontrun{
# setup fastf1 connection with all defaults
setup_fastf1()

# setup with a preexisting conda environment, with a specified name
setup_fastf1("example_conda_env", conda = TRUE)
}
}
