% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dark_f1.R
\name{theme_dark_f1}
\alias{theme_dark_f1}
\title{Dark F1-style Theme for ggplot}
\usage{
theme_dark_f1()
}
\value{
A ggplot object that indicates grand prix, driver, time and selected
color variable.
}
\description{
Theme for all f1dataR plot functions. Mimics Formula 1 style.
}
