% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dark_f1.R
\name{theme_dark_f1}
\alias{theme_dark_f1}
\title{Dark F1-style Theme for ggplot}
\usage{
theme_dark_f1(axis_marks = FALSE)
}
\arguments{
\item{axis_marks}{True or false, whether axis line, ticks and title should
be shown or not. Defaults to false}
}
\value{
A ggplot object that indicates grand prix, driver, time and selected
color variable.
}
\description{
Theme for all f1dataR plot functions. Mimics Formula 1 style.
}
