% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dark_f1.R
\name{theme_dark_f1}
\alias{theme_dark_f1}
\title{Dark F1-style Theme for ggplot}
\usage{
theme_dark_f1()
}
\arguments{
\item{season}{number from 2018 to 2022 (defaults to current season).}

\item{race}{number from 1 to 23 (depending on season selected) and defaults
to most recent.}

\item{driver}{three letter driver code (see load_drivers() for a list)}

\item{color}{argument that indicates which variable to plot overtop the
circuit}
}
\value{
A ggplot object that indicates grand prix, driver, time and selected
color variable.
}
\description{
Theme for all f1dataR plot functions. Mimics Formula 1 style.
}
